KMeans based cluster tree
I implemented a tree based on KMeans clustering. Contrary to the
other trees implmeneted here, the KMeans based tree needs all points
for setup. The KMeans clustering is done with the ParallelKMeans.jl
package.
The depth of the tree can be controlled with the minimal number of
points in a cluster 'nmin' or the maximum level 'maxlevel'.